refactor(server): split god files into focused sub-flow modules
Split profile_need_edit.go (1013 lines) into 6 files:
- profile_need_edit.go (redirect entry point)
- profile_need_edit_location.go
- profile_need_edit_categories.go
- profile_need_edit_story.go
- profile_need_edit_documents.go
- profile_need_edit_review.go

Split onboarding.go (1612 lines) into 7 files:
- onboarding.go (core routing, welcome, shared helpers)
- onboarding_location.go
- onboarding_categories.go
- onboarding_story.go
- onboarding_documents.go
- onboarding_review.go
- onboarding_donor.go

No logic changes — purely mechanical code relocation for navigability.
